+++ Sandswept/Items/Greens/UniversalVIPPass.cs
@@ -9,9 +9,9 @@ namespace Sandswept.Items.Greens
 
         public override string ItemLangTokenName => "VIP_PASS";
 
-        public override string ItemPickupDesc => "Category chests have a chance to drop multiple items.";
+        public override string ItemPickupDesc => "Category chests have a chance to drop an extra item.";
 
-        public override string ItemFullDescription => $"$suCategory chests$se have a $su{chance}%$se chance of dropping $su{baseExtraItems}$se $ss(+{stackExtraItems} per stack)$se $suextra items$se.".AutoFormat();
+        public override string ItemFullDescription => $"$suCategory chests$se have a $su{baseChance * 100f}%$se $ss(+{stackChance * 100f}% per stack)$se chance of dropping $su{extraItems}$se $suextra item$se.".AutoFormat();
 
         public override string ItemLore => "\"I'm sorry, sir, this is a restricted area. We can't allow you in.\"\r\n\r\n\"Special orders from UES. It would be in your best interests to make an exception.\"\r\n\r\n\"I'm afraid I can't do that. We received specific instructions from two of our guests not to let in anyone from UES. It'd be against policy to betray their trust.\"\r\n\r\n\"Those two 'guests' have stolen from the UESC. You're harboring criminals. If you do not let us in, you will be obstructing justice in violation of interplanetary law.\"\r\n\r\n\"The UESC does not have legal jurisdiction over Pluto, sir. We're under no obligation to let you in. If you don't vacate the premises, I'll be forced to call security, and make no mistake, our security is the best of the best.\"\r\n\r\n...\r\n\r\n\"...is that...?\"\r\n\r\n\"Yes. Universal. This is serious business. I'll ask one more time: let us through.\"\r\n\r\n\"...right away, sir.\"\r\n";
 
@@ -21,14 +21,17 @@ namespace Sandswept.Items.Greens
 
         public override Sprite ItemIcon => Main.MainAssets.LoadAsset<Sprite>("UniVIPIcon.png");
 
-        [ConfigField("Chance", "", 50f)]
-        public static float chance;
+        [ConfigField("Base Chance", "Decimal.", 0.4f)]
+        public static float baseChance;
 
-        [ConfigField("Base Extra Items", "", 1)]
-        public static int baseExtraItems;
+        [ConfigField("Stack Chance", "", 0.4f)]
+        public static float stackChance;
 
-        [ConfigField("Stack Extra Items", "", 1)]
-        public static int stackExtraItems;
+        [ConfigField("Extra Items", "", 1)]
+        public static int extraItems;
+
+        [ConfigField("mdmpDebug", "This will break your save file! Only enable this if you know what you are doing!", false)]
+        public static bool dumbassThingLmao;
 
         public static GameObject vfx;
 
@@ -62,12 +65,12 @@ namespace Sandswept.Items.Greens
             transform.Find("ColoredLightShaftsBalance").GetComponent<ParticleSystemRenderer>().material = newColoredLightShaftMat;
 
             var coloredDustBalance = transform.Find("ColoredDustBalance").GetComponent<ParticleSystemRenderer>();
-
+            var tex = dumbassThingLmao ? Main.hifuSandswept.LoadAsset<Texture2D>("texPawMask.png") : Paths.Texture2D.texSpark1Mask;
             var newColoredDustBalanceMat = new Material(Paths.Material.matChanceShrineDollEffect);
             newColoredDustBalanceMat.SetColor("_TintColor", pink);
             newColoredDustBalanceMat.SetTexture("_RemapTex", Paths.Texture2D.texRampAreaIndicator);
             newColoredDustBalanceMat.SetFloat("_Boost", 12f);
-            newColoredDustBalanceMat.SetTexture("_MainTex", Paths.Texture2D.texSpark1Mask);
+            newColoredDustBalanceMat.SetTexture("_MainTex", tex);
 
             VFXUtils.MultiplyScale(vfx, 3f);
             VFXUtils.MultiplyDuration(vfx, 1.5f);
@@ -80,36 +83,7 @@ namespace Sandswept.Items.Greens
 
         public override void Hooks()
         {
-            // SceneDirector.onPostPopulateSceneServer += SceneDirector_onPostPopulateSceneServer;
             On.RoR2.GlobalEventManager.OnInteractionBegin += GlobalEventManager_OnInteractionBegin;
-            // Run.onRunDestroyGlobal += Run_onRunDestroyGlobal;
-        }
-
-        private void Run_onRunDestroyGlobal(Run run)
-        {
-            totalCategoryChests = 0;
-        }
-
-        public static int totalCategoryChests = 0;
-
-        private void SceneDirector_onPostPopulateSceneServer(SceneDirector sd)
-        {
-            int currentStageCategoryChests = 0;
-            var purchaseInteractions = GameObject.FindObjectsOfType<PurchaseInteraction>();
-            for (int i = 0; i < purchaseInteractions.Length; i++)
-            {
-                var purchaseInteraction = purchaseInteractions[i];
-                if (purchaseInteraction.displayNameToken.ToLower().Contains("category") || purchaseInteraction.contextToken.ToLower().Contains("category"))
-                {
-                    totalCategoryChests++;
-                    currentStageCategoryChests++;
-                }
-            }
-
-            Main.ModLogger.LogError("Stage: " + SceneManager.GetActiveScene().name);
-            Main.ModLogger.LogError("Category chests on stage: " + currentStageCategoryChests);
-            Main.ModLogger.LogError("Category chests this run: " + totalCategoryChests);
-            Main.ModLogger.LogError("You would get this many items on average with a chance of " + chance + "%: " + (totalCategoryChests * (chance / 100f)));
         }
 
         private void GlobalEventManager_OnInteractionBegin(On.RoR2.GlobalEventManager.orig_OnInteractionBegin orig, GlobalEventManager self, Interactor interactor, IInteractable interactable, GameObject interactableObject)
@@ -138,14 +112,15 @@ namespace Sandswept.Items.Greens
                         {
                             var stack = GetCount(interactorBody);
                             var scale = 0.5f + Run.instance.participatingPlayerCount * 0.5f;
+
+                            var chance = MathHelpers.InverseHyperbolicScaling(baseChance, stackChance, 1f, stack) * 100f;
+
                             if (stack > 0 && Util.CheckRoll(chance / scale))
                             {
                                 var isCategoryChestFinal = interactableObject.name.ToLower().Contains("category") || isCategoryChest;
                                 if (isCategoryChestFinal)
                                 {
-                                    var extraItemCount = baseExtraItems + stackExtraItems * (stack - 1);
-
-                                    chestBehavior.dropCount = 1 + extraItemCount;
+                                    chestBehavior.dropCount = 1 + extraItems;
 
                                     Util.PlaySound("Play_UI_commandHUD_select", chestBehavior.gameObject);
                                     Util.PlaySound("Play_UI_commandHUD_select", chestBehavior.gameObject);
@@ -158,10 +133,12 @@ namespace Sandswept.Items.Greens
                                         color = pink
                                     }, true);
 
-                                    if (Random.Range(0f, 100f) >= 96f)
+                                    Util.PlaySound("Play_UI_obj_casinoChest_open", interactableObject);
+                                    Util.PlaySound("Play_ui_obj_lunarPool_activate", interactableObject);
+
+                                    if (dumbassThingLmao)
                                     {
-                                        Chat.AddMessage("<style=cIsDamage>Developer 1</style>: Universal VIP Paws :3 x3 OwO UwU :3 <3");
-                                        Chat.AddMessage("<style=cIsUtility>Developer 2</style>: What???");
+                                        Chat.AddMessage("<color=#DC4C7B>Universal VIP Paws :3 x3 OwO UwU :3 <3</color>");
                                     }
                                 }
                             }